requirements.txt
discord.py==1.2.5
pytz
pymongo
python-dotenv